Implement RADIUS salt en- and decryption
This commit implement utility function for encrypting and decrypting
salt encrypted RADIUS Attributes.

The data for the testcases were generated with the freeradius tool
"radclient".

The encryption and decryption functions have asymmetric parameters. The
salt for the decryption function is read from the encoded attribute
itself.

Generation of the salt for the encryption cannot be automated. RFC 2828
mandates that the salt of the encrypted Tunnel-Password attribute has
the most significant bit set to 1. The Draft
draft-ietf-radius-saltencrypt-00 requires that the most significant bit
is set to 1 for responses type packets and to 0 for request type
packets. Since this information is not available to the encrypt function
the caller has to pass a salt explicitly.

RFC 2828 suggest to pad the salt encrypted value with zeros, which is
done in this implementation. This differs from the Freeradius
implementation (which was used as a reference) where the padding was
filled with random data.